PM-24565: Syncronize token refreshes to avoid duplicate requests
