+++ network/src/main/kotlin/com/bitwarden/network/BitwardenServiceClientImpl.kt
@@ -2,8 +2,7 @@ package com.bitwarden.network
 
 import com.bitwarden.annotation.OmitFromCoverage
 import com.bitwarden.core.data.serializer.ZonedDateTimeSerializer
-import com.bitwarden.network.authenticator.RefreshAuthenticator
-import com.bitwarden.network.interceptor.AuthTokenInterceptor
+import com.bitwarden.network.interceptor.AuthTokenManager
 import com.bitwarden.network.interceptor.BaseUrlInterceptors
 import com.bitwarden.network.interceptor.HeadersInterceptor
 import com.bitwarden.network.model.BitwardenServiceClientConfig
@@ -52,8 +51,7 @@ internal class BitwardenServiceClientImpl(
     private val bitwardenServiceClientConfig: BitwardenServiceClientConfig,
 ) : BitwardenServiceClient {
 
-    private val refreshAuthenticator: RefreshAuthenticator = RefreshAuthenticator()
-    private val authTokenInterceptor: AuthTokenInterceptor = AuthTokenInterceptor(
+    private val authTokenManager: AuthTokenManager = AuthTokenManager(
         clock = bitwardenServiceClientConfig.clock,
         authTokenProvider = bitwardenServiceClientConfig.authTokenProvider,
     )
@@ -75,7 +73,7 @@ internal class BitwardenServiceClientImpl(
     }
     private val retrofits: Retrofits by lazy {
         RetrofitsImpl(
-            authTokenInterceptor = authTokenInterceptor,
+            authTokenManager = authTokenManager,
             baseUrlInterceptors = BaseUrlInterceptors(
                 baseUrlsProvider = bitwardenServiceClientConfig.baseUrlsProvider,
             ),
@@ -84,7 +82,6 @@ internal class BitwardenServiceClientImpl(
                 clientName = bitwardenServiceClientConfig.clientData.clientName,
                 clientVersion = bitwardenServiceClientConfig.clientData.clientVersion,
             ),
-            refreshAuthenticator = refreshAuthenticator,
             logHttpBody = bitwardenServiceClientConfig.enableHttpBodyLogging,
             certificateProvider = bitwardenServiceClientConfig.certificateProvider,
             json = clientJson,
@@ -206,7 +203,6 @@ internal class BitwardenServiceClientImpl(
     }
 
     override fun setRefreshTokenProvider(refreshTokenProvider: RefreshTokenProvider?) {
-        refreshAuthenticator.refreshTokenProvider = refreshTokenProvider
-        authTokenInterceptor.refreshTokenProvider = refreshTokenProvider
+        authTokenManager.refreshTokenProvider = refreshTokenProvider
     }
 }
